Added more transitions, more tests, and added listeners
Added Cut transition
Added more tests
Added tally change via TCP api
Added listener for tally change of states

+++ src/main/java/uk/co/flakeynetworks/vmix/status/Audio.java
@@ -7,7 +7,7 @@ import org.simpleframework.xml.Root;
 public class Audio {
 
     @Attribute
-    private int volume;
+    private double volume;
 
     @Attribute
     private boolean muted;
